Got SSL working, change user identification
I also converted some spaces to tabs since it was looking weird in my editor, and fixed the bug of the program not exiting for various reasons.  Right now I've changed the user identification model so that instead of the user sending their public key, we look for their nickname in the certificate.  The idea would be trust the certificate authorities to verify their nickname.  The good thing about this is that it is relying on the SSL authentication.  This way users won't be able to pretend to be other users.  This requires that we create the certificates correctly though.  Additionally instead of having a server certificate and a bunch of client certificates each user always uses the same key and certificate.